+++ tests/Integration/Webhook/ServerTest.php
@@ -20,7 +20,7 @@ class ServerTest extends \PHPUnit_Framework_TestCase
 
     public function setUp()
     {
-        $this->guzzleClient = new Client('http://localhost:8000');
+        $this->guzzleClient = new Client('http://localhost:8999');
     }
 
     /**
@@ -33,7 +33,7 @@ class ServerTest extends \PHPUnit_Framework_TestCase
         $testCase,
         $testHeaders
     ) {
-        $process = new Process('php -S localhost:8000', __DIR__.'/../../resources');
+        $process = new Process('php -S localhost:8999', __DIR__.'/../../resources');
         $process->start();
         sleep(1);
         $signature = sha1($request.ServerMock::PROJECT_SECRET_KEY);
@@ -56,7 +56,7 @@ class ServerTest extends \PHPUnit_Framework_TestCase
         static::assertSame(Version::getVersion(), (string) $response->getHeader('x-xsolla-sdk'));
         static::assertArrayHasKey('content-type', $response->getHeaders());
         if (204 === $response->getStatusCode()) {
-            static::assertSame('text/plain;charset=UTF-8', (string) $response->getHeader('content-type'));
+            static::assertSame('text/plain', (string) $response->getHeader('content-type'));
         } else {
             static::assertSame('application/json', (string) $response->getHeader('content-type'));
         }